Добалена возвможность в балы в табличек Эффективности менеджера вписывать числа с запетой, исправлена табилица с статистикой по звонкам на странице очетов у Контролера - убраны со статистики ватсапы и письма

+++ Data/Entities/Calls/PerformanceChart.cs
@@ -5,14 +5,14 @@ namespace Data.Entities.Calls
     public class PerformanceChart : Entity
     {
         public int ManagerId { get; set; }
-        public int NumberPlan_DevelopmentCalls { get; set; }
-        public int NumberPlan_YourShifts { get; set; }
-        public int NumberPlan_SubstitutionShifts { get; set; }
-        public int ShiftPlan_DevelopmentCalls { get; set; }
-        public int ShiftPlan_YourShifts { get; set; }
-        public int ShiftPlan_SubstitutionShifts { get; set; }
-        public int Balls_DevelopmentCalls { get; set; }
-        public int Balls_YourShifts { get; set; }
-        public int Balls_SubstitutionShifts { get; set; }
+        public double NumberPlan_DevelopmentCalls { get; set; }
+        public double NumberPlan_YourShifts { get; set; }
+        public double NumberPlan_SubstitutionShifts { get; set; }
+        public double ShiftPlan_DevelopmentCalls { get; set; }
+        public double ShiftPlan_YourShifts { get; set; }
+        public double ShiftPlan_SubstitutionShifts { get; set; }
+        public double Balls_DevelopmentCalls { get; set; }
+        public double Balls_YourShifts { get; set; }
+        public double Balls_SubstitutionShifts { get; set; }
     }
 }
